Refactor changes into dedicated modules/fns
